Fix SIGINT/SIGTERM failing to stop crawling
Close #6
